Backend/djangoProject/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bleach==5.0.0
cffi==1.15.1
cryptography==38.0.1
Django==4.0.5
django-cors-headers==3.13.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.2.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
